add edit for admin

+++ src/controller/bookController.js
@@ -1,6 +1,9 @@
 const createError = require("../Utilities/createError");
 const prisma = require("../model/prisma");
-const { makeBookingSchema } = require("../validators/bookValidator");
+const {
+    makeBookingSchema,
+    makeBookingSchemaForEdit,
+} = require("../validators/bookValidator");
 const { bookingItemScehema } = require("../validators/bookValidator");
 
 exports.createBooking = async (req, res, next) => {
@@ -12,11 +15,23 @@ exports.createBooking = async (req, res, next) => {
         value.userId = req.user.id;
         const { serviceId, ...bookingDetail } = value;
 
+        bookingDetail.hairStylistId = +bookingDetail.hairStylistId;
+
+        console.log(bookingDetail);
+
         const findBooked = await prisma.booking.findFirst({
             where: bookingDetail,
         });
 
-        if (!findBooked) {
+        console.log(findBooked);
+
+        if (findBooked) {
+            return res
+                .status(400)
+                .json({ message: "วันและเวลาดังกล่าวได้ภูกจองไว้แล้ว" });
+        }
+
+        if (!findBooked && req.user.role !== "ADMIN") {
             const booking = await prisma.booking.create({
                 data: bookingDetail,
             });
@@ -29,6 +44,7 @@ exports.createBooking = async (req, res, next) => {
             });
             return res.status(200).json({ message: "Create success" });
         }
+        return res.status(400).json({ message: "ADMIN CANT BOOK!!" });
     } catch (error) {
         next(error);
     }
@@ -58,6 +74,9 @@ exports.getBooking = async (req, res, next) => {
                 bookTime: {
                     select: { bookTime: true },
                 },
+            },
+            orderBy: {
+                bookDate: "asc",
             },
         });
         res.status(200).json(targetBooking);
@@ -97,7 +116,7 @@ exports.getBookingForAdmin = async (req, res, next) => {
                 },
             },
             orderBy: {
-                bookDate: "asc",
+                bookTimeId: "desc",
             },
         });
         res.status(200).json(targetBooking);
@@ -133,20 +152,26 @@ exports.deleteBooking = async (req, res, next) => {
         }
         const bookedItem = await prisma.booking.findFirst({
             where: {
-                id: value.bookId,
+                id: value.bookedId,
                 userId: req.user.id,
             },
         });
+        console.log(bookedItem.id);
         if (!bookedItem) {
             return next(createError("can not delete booking "));
         }
 
-        await prisma.booking.delete({
-            where: {
-                id: bookedItem.id,
-            },
-        });
-        res.status(200).json({ message: "delete success" });
+        if (bookedItem.status === "ACCEPTED") {
+            return res.status(400).json({ message: "can not delete booked" });
+        }
+        if (bookedItem.status !== "ACCEPTED") {
+            await prisma.booking.delete({
+                where: {
+                    id: bookedItem.id,
+                },
+            });
+            return res.status(200).json({ message: "delete success" });
+        }
     } catch (error) {
         next(error);
     }
@@ -155,16 +180,74 @@ exports.deleteBooking = async (req, res, next) => {
 exports.editBookingForUser = async (req, res, next) => {
     try {
         const { value, error } = bookingItemScehema.validate(req.params);
+        const { ...validateObj } = makeBookingSchemaForEdit().validate(
+            req.body
+        );
+
         if (error) {
             return next(createError("Can not edit this booking"));
         }
+
+        if (validateObj.error) {
+            return next(createError("โปรดกรอกข้อมูลให้ครบถ้วน"), 400);
+        }
+
         const bookedItem = await prisma.booking.findFirst({
             where: {
-                id: value.bookId,
+                id: value.bookedId,
                 userId: req.user.id,
             },
         });
-        res.status(200).json(bookedItem);
+
+        if (!bookedItem) {
+            return next(createError("can not edit booking "));
+        }
+
+        if (
+            bookedItem.status === "ACCEPTED" ||
+            bookedItem.status === "DENINED"
+        ) {
+            return res.status(400).json({ message: "can not edit booked" });
+        }
+
+        console.log(validateObj.value);
+
+        const newBookedItem = await prisma.booking.update({
+            data: validateObj.value,
+            where: bookedItem,
+        });
+
+        res.status(200).json(newBookedItem);
+    } catch (error) {
+        next(error);
+    }
+};
+
+exports.editStatusBookingForAdmin = async (req, res, next) => {
+    try {
+        const { value, error } = bookingItemScehema.validate(req.params);
+        const { status } = req.body;
+        if (error) {
+            return next(error);
+        }
+        const bookedItem = await prisma.booking.findFirst({
+            where: {
+                id: value.bookedId,
+                hairStylistId: req.user.id,
+            },
+        });
+        console.log(bookedItem);
+
+        if (!bookedItem) {
+            return next(createError("can not edit booking "));
+        }
+
+        await prisma.booking.update({
+            data: { status },
+            where: bookedItem,
+        });
+
+        res.status(200).json({ message: "Edit Success" });
     } catch (error) {
         next(error);
     }
